LayoutTests/fast/loader/subframe-navigate-during-main-frame-load.html

<html>
<head>
<script>

function runTest()
{
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.dumpBackForwardList();
        layoutTestController.waitUntilDone();
    }

    // Location changes need to happen outside the onload handler to generate history entries.
    setTimeout(function() {location='resources/subframe-navigate-during-main-frame-load2.html';}, 0);
}

</script>
</head>

<body onload="runTest();">
You shouldn't even see this page, it navigates away immediately!
</body>

</html>
