LayoutTests/fast/loader/subframe-navigate-during-main-frame-load-expected.txt

This test navigates an iframe while it's onload handler is still running.
The iframe navigation should *not* result in a second back-history item being created for the top level navigation.



============== Back Forward List ==============
        (file test):fast/loader/subframe-navigate-during-main-frame-load.html  **nav target**
curr->  (file test):fast/loader/resources/subframe-navigate-during-main-frame-load2.html  **nav target**
            data:text/html,%3Cbody%20onload=%22layoutTestController.notifyDone();%22%3E%3C/body%3E (in frame "subframe")
===============================================
